Add related_posts option RELATED_POSTS_SKIP_SAME_CATEGORY
Posts in the same category may sometimes (depending on the
overall site design) have a lot of tags in common. Thus the
`article.related_posts` holds only the posts that are tightly
(perhaps too tightly) related to the original post, and no
posts from other categories, since those, with only one or two
tags in common, can't as effectively compete.