Refactored filetime_from_git and added GIT_FILETIME_FOLLOW option.
The filetime_from_git plugin relies on a now old version of gitpython.
However its not so easy as updating to the latest version as your
gitpython version must match your distros git binary for everything to
work.

This refactor restructures the plugin to provide some isolation from
this issue by wrapping the gitpython module in a simplified wrapper to
handle this difference.

I have also added the GIT_FILETIME_FOLLOW option which allows the
filetime module to follow files across renames.