Updated plugin to enhance privacy with video embeds.
This plugin is a conceptual replication of the Electronic Frontier
Foundation's MyTube plugin, which pre-caches thumbnails of embedded
YouTube videos so that users are not tracked by Google until they
explicitly "opt-in" by clicking on the thumbnail (at which point the
thumbnail is replaced by the video embed iframe).

The plugin now supports both YouTube and Vimeo, and is now easier to extend for other video services.