Add better-tables pelican plugin - for restoring sanity to tables
Simply, this plugin removes the extraneous elements and attributes that
rst->html generates. Without this cruft, the tables pass HTML5 validation and
make life /much/ easier when styling with CSS.

The real fix is to get rst2html5 merged, but sadly that is beyond both my
skillset and time/effort availablity.