Test all plugins with runtests.py script
This script allow to run the tests in multiple environments with tox.
Find automatically plugins with tests and with a custom tox.ini file.
Custom tox config are run independently.
All over tests are run in the same run.
Read args to tests only certain plugins

A requirements.txt file can be added to the plugin dir to install the
packages for the test run.

Write tests results to a csv file