Add GIT_FILETIME_ALWAYS_ADD_MODIFIED setting.
The plugin usually always sets the "modified" field of a content piece,
even if it has not actually been modified. This setting allows the user
to disable that behaviour so that only content pieces which have been
modified get a "Modified" date.

To ensure backward compatibility, this setting is set to "True" by
default.

Signed-off-by: Tilman Blumenbach <tilman+git@ax86.net>