[extract_toc] Fix pandoc_reader Issue
Fixes #582

toc_extract assumes that if it can import pandoc_reader that
pandoc_reader is configured. This is not always the case and can cause
an error as pandoc_reader requires pelicanonf parameter.

This fix adds a plugin check to make sure that pandoc_reader was meant
to be called by having it as a configured setting.