Added support for attached images
Does not assume file path is absolute, and does not require image to be copied yet.
If detects {attach} in filename-only then it will look relative to the current data file (.md for example) for reading properties. Otherwise maintains the way it works now, have tested with data for example and the conditional only triggers when {attach} is seen, all else is as per existing.