liquid_tags/youtube: embed linked thumbnail
Adding <iframe> adds 1.2MB of javascript to webpage. And tracking
cookies.

Instead we can add thumbnail of video and link it to YouTube page. So
user can click to watch movie and is not tracked on our site.

YOUTUBE_THUMB_ONLY liquid config flag enables this mode
YOUTUBE_THUMB_SIZE allows to select one of YT thumbnails - there are
four of them defined:

- maxres: 1280x720
- sd:      640x360
- hq:      480x360
- mq:      320x180

'sd' one is used if not defined in LIQUID_CONFIGS variable.