Simplify Webring plugin HTTPS support
After reading that feedparser will remove its internal http client in
following releases (https://github.com/kurtmckee/feedparser/pull/178),
it made sense to stop using it, which in effect removes the need for the
WEBRING_SUPPORT_HTTPS variable.

The Requests library was not used because the requirements were pretty
simple and also to avoid another dependency.