fix(liquid_tags.include_code): do not put code inside figcaption. Add CSS classes
A regression was introduced few patches ago, which placed code inside
the `figcaption` tag. This is not correct.

The original version has this layout

```
figure
  figcaption
  code
```

This patch fixes this issue. It also encloses metadata,

1. title
2. filename
3. line number

inside CSS classes. This let's themes improve the look of the
figcaption. Previously, all the metadata was piled into one `<span>`.