liquid_tags: Registering md preprocessor with register() rather than the deprecated add().
<rant>
The semantics of add() were inverse to the intuition because they were
implicitly exposing the internals of markdown.util.Registry.

E.g. a priority of ">htmlblock" did not translate to "have greater
priority than htmlblock processor", as one would expect. Instead it
means "when the interal processors array is sorted by priority, the
new processor should have a bigger *index* number than htmlblock".
The tricky point here is that the internal array is sorted in reverse.
So ">htmlblock" actually translates to "use a smaller priority than
htmlblock". Madness!

And instead of fixing the semantics to make sense, it was chosen
to deprecate add() and force us to directly supply the priority to
register(). This is also bad, because we have to grep outside our
codebase to determine the proper priority to achieve the desired effect.
Double the madness!
</rant>