Have simple_bookmarks always quote attributes
The regex in `Content._get_intrasite_link_regex()` expects attributes to
be quoted. When `quote_attr_values` has the value `legacy` when
serialising the parsed HTML with html5lib, it ends up omitting the
quotes expected, which breaks intrasite links. Changing this to `always`
ensures that the quotes remain present if simple_bookmarks processes the
page.

I expect this will also fix #1240.